% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP_PROCEDURE.R
\name{BOOTSTRAP_PROCEDURE}
\alias{BOOTSTRAP_PROCEDURE}
\title{Estimate the variances at phase two for the regression coefficients by
using a bootstrap procedure}
\usage{
BOOTSTRAP_PROCEDURE(
  brepsMax,
  phase1_size,
  phase2_size,
  member_phase2,
  DELTA,
  OMEGA,
  Gama_initial,
  d,
  qn,
  ZMat,
  UVec,
  VVec,
  knot,
  degree,
  boundary,
  GENERATE_GRADIENT_HESSIAN,
  GENERATE_PHI
)
}
\arguments{
\item{brepsMax}{The number of bootstrap random samples.}

\item{phase1_size}{A vector consisting of the sample sizes of all
strata at the first phase.}

\item{phase2_size}{A vector consisting of the sample sizes of all
strata at the second phase.}

\item{member_phase2}{A vector consisting of the observations
which are selected to form those at the second phase.}

\item{DELTA}{A matrix that each row indicates the interval-censoring
type of the observation with the sample size by 3. For example,
\verb{(1, 0, 0)} indicates that \code{UVec} is observed but \code{VVec} reaches the
right boundary, \verb{(0, 1, 0)} indicates both \code{UVec} and \code{VVec} are
observed and \verb{(0, 0, 1)} indicates \code{UVec} reaches the left boundary
but \code{VVec} is observed for the observation.}

\item{OMEGA}{A vector that consisting of the weights for each observation
after two-phase sampling. If the two-phase sampling is not used,
the values of \code{OMEGA} are all 1's.}

\item{Gama_initial}{An initial value of \code{Gama}.}

\item{d}{The number of unknown coefficients of the regression parameters.}

\item{qn}{The number of unknown coefficients of the B-spline basis,
that is, \code{df} or \code{degree + length(knot) + 1} (because the intercept
is included).}

\item{ZMat}{ZMat The design matrix corresponding to the regression
parameters with the sample size by \code{d}.}

\item{UVec}{A vector consisting of the left observation visit given
in the interval-censoring design.}

\item{VVec}{A vector consisting of the right observation visit given
in the interval-censoring design.}

\item{knot}{The internal breakpoints (i.e., the breakpoints without two
ending points) that define the spline. The default is \code{NULL}, which
results in a basis for the ordinary polynomial regression. Typical
values are the mean or median for one knot, quantiles for more knots.
See also \code{Boundary.knots}.}

\item{degree}{A non-negative integer degree of the piece-wise polynomial.
The default value is 3 for cubic splines. Zero degree is allowed for
this function, which is the only difference compared with \code{bs}
function in package \code{splines}. Note that the order of a polynomial
equals to the degree plus one.}

\item{boundary}{A 2-dimensional boundary points at which to anchor the
B-spline basis. By default, they are the range of the non-\code{NA} data.
If both \code{knot} and \code{boundary} are supplied, the basis parameters do
not depend on \code{x}. Data can extend beyond \code{boundary}.}

\item{GENERATE_GRADIENT_HESSIAN}{A function numerically calculates the
gradient vector and the Hessian matrix of the given weighted
log-likelihood function under two-phase sampling design.}

\item{GENERATE_PHI}{The function generates the B-spline basis matrix and
the values of a linear combination of B-spline basis function at a
given vector.}
}
\value{
A \code{brepsMax} by \code{length(Gama)} matrix has the bootstrap result.
}
\description{
Estimate the variances at phase two for the regression coefficients by
using a bootstrap procedure
}
\references{
\url{https://onlinelibrary.wiley.com/doi/10.1111/sjos.12152}
}
