% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstFunctions.R
\name{FirstFunctions}
\alias{FirstFunctions}
\title{This is the title for our FirstFunctions, Returns the sum of two numbers}
\usage{
FirstFunctions(a, b)
}
\arguments{
\item{a}{Explain the parameter a, say, Number One}

\item{b}{Explain the parameter b, say, Number Two}
}
\value{
Explain what is the result, say, Sum of two numbers
}
\description{
This is the title for our FirstFunctions, Returns the sum of two numbers
}
\examples{
library(VarEst2Phase)
FirstFunctions(1, 2)
}
