% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR.R
\name{GR}
\alias{GR}
\title{Estimate both the coefficients of the regression parameters and the
coefficients of the B-spline basis functions and estimate the variance
of the estimated coefficents of the regression parameters by using
Generalized Rosen (GR) algorithm}
\usage{
GR(
  gama.ini,
  active.set,
  d,
  qn,
  ZMat,
  UVec,
  VVec,
  knot,
  degree,
  boundary,
  DELTA,
  OMEGA,
  GENERATE_GRADIENT_HESSIAN,
  GENERATE_PHI
)
}
\arguments{
\item{gama.ini}{An initial value of \code{Gama}.}

\item{active.set}{A vector includes an active set corresponding to
\code{gama.ini}.}

\item{d}{The number of unknown coefficients of the regression parameters.}

\item{qn}{The number of unknown coefficients of the B-spline basis,
that is, \code{df} or \code{degree + length(knot) + 1} (because the intercept
is included).}

\item{ZMat}{The design matrix corresponding to the regression
parameters with the sample size by \code{d}.}

\item{UVec}{A vector consisting of the left observation visit given
in the interval-censoring design.}

\item{VVec}{A vector consisting of the right observation visit given
in the interval-censoring design.}

\item{knot}{The internal breakpoints (i.e., the breakpoints without two
ending points) that define the spline. The default is \code{NULL}, which
results in a basis for the ordinary polynomial regression. Typical
values are the mean or median for one knot, quantiles for more knots.
See also \code{Boundary.knots}.}

\item{degree}{A non-negative integer degree of the piece-wise polynomial.
The default value is 3 for cubic splines. Zero degree is allowed for
this function, which is the only difference compared with \code{bs}
function in package \code{splines}. Note that the order of a polynomial
equals to the degree plus one.}

\item{boundary}{A 2-dimensional boundary points at which to anchor the
B-spline basis. By default, they are the range of the non-\code{NA} data.
If both \code{knot} and \code{boundary} are supplied, the basis parameters do
not depend on \code{x}. Data can extend beyond \code{boundary}.}

\item{DELTA}{A matrix that each row indicates the interval-censoring
type of the observation with the sample size by 3. For example,
\verb{(1, 0, 0)} indicates that \code{UVec} is observed but \code{VVec} reaches the
right boundary, \verb{(0, 1, 0)} indicates both \code{UVec} and \code{VVec} are
observed and \verb{(0, 0, 1)} indicates \code{UVec} reaches the left boundary
but \code{VVec} is observed for the observation.}

\item{OMEGA}{A vector that consisting of the weights for each observation
after two-phase sampling. If the two-phase sampling is not used,
the values of \code{OMEGA} are all 1's.}

\item{GENERATE_GRADIENT_HESSIAN}{A function numerically calculates the
gradient vector and the Hessian matrix of the given weighted
log-likelihood function under two-phase sampling design.}

\item{GENERATE_PHI}{The function generates the B-spline basis matrix and
the values of a linear combination of B-spline basis function at a
given vector.}
}
\value{
A list includes: a vector consists the estimated coefficients of
the regression parameters, an error message corresponding to the
Hessian matrix, a status message corresponding to the number of
replications of updating \code{lamda} or \code{gama} during the calculation,
and an estimated covariance matrix of the regression coefficients.
}
\description{
Estimate both the coefficients of the regression parameters and the
coefficients of the B-spline basis functions and estimate the variance
of the estimated coefficents of the regression parameters by using
Generalized Rosen (GR) algorithm
}
